+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.25.0
+
+- **BREAKING CHANGE**: rename `FlowResult` to `ProcessResult`
+- **BREAKING CHANGE**: rename `dispatchEvent()` to `runProcess()`
+- **FEAT**: use horda_core 0.17.0
+
 ## 0.24.0
 
 - **FEAT**: add `RemoteMessageExtensions` on `BuildContext` with methods for entity communication: